new try

+++ arraylist.c
@@ -38,7 +38,7 @@ void push(ArrayList * l, void * data, int i){
   }
   if(i<l->size)
   {
-    for(size_t k=l->size; k>i; k--){
+    for(int k=l->size; k>i; k--){
       l->data[k] = l->data[k-1];
     }
     l->data[i]=data;
